refactor: 预分离 LInt.h
